+++ molviewspec/molviewspec/__init__.py
@@ -1,6 +1,6 @@
 """MolViewSpec"""
 
-__version__ = "1.4.0"
+__version__ = "1.5.0"
 
 from molviewspec.builder import create_builder
 from molviewspec.nodes import (