+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file, following the suggestions of [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+
+## [v1.5.0] - 2025-04-09
+
 - Fix missing tooltip default value for `BoxParams`
 - Add simple widgets for Jupyter & Colab (`builder.molstar_notebook()`) and Streamlit (`builder.molstar_streamlit()`)